requirements.txt
Pillow>=9.0.0
torch>=2.0.0
numpy>=1.24.0